requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
build==1.4.0
certifi==2026.2.25
charset-normalizer==3.4.4
chromadb==1.5.1
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
durationpy==0.10
et_xmlfile==2.0.0
filelock==3.24.3
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.2.0
googleapis-common-protos==1.72.0
grpcio==1.78.1
h11==0.16.0
hf-xet==1.3.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kubernetes==35.0.0
markdown-it-py==4.0.0
matplotlib==3.10.8
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
narwhals==2.18.0
numpy==2.4.2
oauthlib==3.3.1
ollama==0.6.1
onnxruntime==1.24.2
openpyxl==3.1.5
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.7
overrides==7.7.0
packaging==26.0
pandas==3.0.1
pillow==12.1.1
plotly==6.6.0
posthog==5.4.0
protobuf==6.33.5
pybase64==1.4.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.19
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.3
rpds-py==0.30.0
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
tenacity==9.1.4
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
zipp==3.23.0
